% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R, R/RDA_crossvalidation.R
\name{alpha_gamma_crossFit}
\alias{alpha_gamma_crossFit}
\title{alpha_gamma_crossFit}
\usage{
alpha_gamma_crossFit(data_set)

alpha_gamma_crossFit(data_set)
}
\arguments{
\item{data_set}{to be trained on for cross validation}
}
\value{
alpha, gamma

alpha, gamma
}
\description{
cross validates for the best alpha and gamma

cross validates for the best alpha and gamma
#'@param data_set to be trained on for cross validation
}
