% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basis_expansion.R
\name{basis_exp}
\alias{basis_exp}
\title{basis_exp}
\usage{
basis_exp(type = "id")
}
\arguments{
\item{type}{One of the following strings: \itemize{\item "id"; \item "quad"; \item "cube"; \item "sqrt"; \item "log"; \item "abs"}}
}
\value{
A function that expands a vector or data-frame by the specified type
}
\description{
A function for expanding data-frames or vectors in aspecific way. Used in @seealso PDA to expand the data into a higher
dimensional space
}
\examples{
h <- basis_exp("quad")
h(c(2, 3))
[1] 2 3 4 6 9
}
