% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R, R/plot_functions.R
\name{calc_error}
\alias{calc_error}
\title{calc_error}
\usage{
calc_error(set, name)

calc_error(set, name)
}
\arguments{
\item{set}{A object of class 'data_set'}

\item{name}{the functionname of the function in set}

\item{data}{Dataframe of Parameters for all Observations}

\item{results}{correct classes}

\item{f}{classification function}
}
\value{
total error rate

A list with 3 entrys:
}
\description{
calculates the total error rate of of a classifaction function on a dataset

Calculates the following percentages for a function of a data_set:
an observation of a class 'A' classified to a class 'B'
class 'A' classified to 'A'
class 'A' not classified to 'A' 
classified to class 'A' actually was class 'B'
classified to class 'A' and also was 'A'
classified to class 'A' but was not 'A'
total missclassification of the data
}
