% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyplot.R
\name{classify_app}
\alias{classify_app}
\title{classify_app}
\usage{
classify_app()
}
\value{
Nothing
}
\description{
A shiny application for classifying data with a graphic interface. Use the "Load Test" option to 
load in a R6 data_set object by its name or generate a random test using the options on the 
left and the paramters for the std in the middle. For classifying use the column on the right 
and chose the classifier and potentially and further parameters. To view the classification 
plot go to the second tab and to see the error plot go to the thrid tab. In both these the
image can be saved with the options  at the bottom. The images will be saved as .png
}
