% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_cov.R
\name{emp_cov}
\alias{emp_cov}
\title{emp_cov}
\usage{
emp_cov(x, y)
}
\arguments{
\item{x}{A numeric Vector}

\item{y}{A numeric Vector of same length as x}
}
\value{
The empirical Covariance of x and y
}
\description{
emp_cov
}
\examples{
emp_cov(c(1, 2, 3), c(1, 2, 3))
emp_cov(c(1, 2, 3), c(3, 2, 1))
}
