% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{is.data_set}
\alias{is.data_set}
\title{is.data_set}
\usage{
is.data_set(set)
}
\arguments{
\item{set}{An R Object}
}
\value{
TRUE if set is a data_set, else FALSE
}
\description{
Checks if the input is an Object of class data_set.
}
\examples{
is.data_set(NULL)

#FALSE
}
