% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moore_penrose.R
\name{lambda_plus}
\alias{lambda_plus}
\title{lambda_plus}
\usage{
lambda_plus(vec)
}
\arguments{
\item{vec}{A Vector}
}
\value{
the Lambda-Plus vector of the given Vector
}
\description{
lambda_plus
}
\examples{
lambda_plus(c(1,4,2))
lambda_plus(c(1, 0, 4, 0))
}
