% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_funs.R
\name{LDA}
\alias{LDA}
\title{LDA}
\usage{
LDA(set)
}
\arguments{
\item{set}{A R6 data_set object initialized with make_set. @seealso make_set}
}
\value{
Returns a list with the name of the created LDA function in the given set in the first entry and the actual classification
function in the second entry and saves it in the input set.
}
\description{
The LDA classification function as described in Hastie et al. "The Elements Statistical Learning" (2009)
}
\examples{
test <- make_testset()
func_name <- LDA(test)[['name']]
}
