% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{maincomponent_analysis}
\alias{maincomponent_analysis}
\title{maincomponent_analysis}
\usage{
maincomponent_analysis(set)
}
\arguments{
\item{set}{a data_set}
}
\value{
matrix whose columns contain the eigenvectors of set$sigma_bet.
The vectors are normalized to unit length and sorted by decreasing eigenvalues.
}
\description{
maincomponent_analysis
}
