% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{make_2D_plot}
\alias{make_2D_plot}
\title{make_2D_plot}
\usage{
make_2D_plot(set, func_name = FALSE, ppu = 5, project = TRUE,
  bg = TRUE)
}
\arguments{
\item{set}{a data_set}

\item{func_name}{Functionname of the data_set function to plot as string. If equals FALSE
just the first two columns or the the columns defined with project will be plotted as points in a plane.}

\item{ppu}{points per unit of the backgoundgrid as integer}

\item{project}{logical or vector with a selection of two data columns.If logical: defines if maincomponent analysis should be used.}

\item{bg}{logical, defines if a background grid should be generated}
}
\value{
A 2 dimensional plot of the data_set with the function as
}
\description{
Takes a data_set and creates a 2 dimensional plot
}
\examples{
test <- make_test()
set <- make_set(test, by='class')
make_2D_plot(set, project = c(1,1), project=FALSE)

\dontrun{
make_2D_plot(set,'LDA_1', ppu = 3)
make_2D_plot(set,'LDA_1', project = c('a','b'))
}
}
