% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{make_projection}
\alias{make_projection}
\title{make_projection}
\usage{
make_projection(set, dim = 2)
}
\arguments{
\item{set}{a data_set}

\item{dim}{dimension of target space}
}
\value{
A list with a function that reduces to dim via maincomponent analysis and one that projects back by setting all other dimensions to zero.
}
\description{
make_projection
}
