% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{make_test}
\alias{make_test}
\title{make_test}
\usage{
make_test(ninputs = 100, nparam = 2, nclasses = 2, sigma = 0.8,
  cube = c(-5, 5))
}
\arguments{
\item{ninputs}{Number of generated Observations per Class}

\item{nparam}{Number of generated parameter columns}

\item{nclasses}{Number of Classes to create.}

\item{cube}{Vector of length 2 with lower and upper boundaries for the parameter centers}

\item{simga}{Numeric Vector of sigma Values for the gaussian distribution togenerate the Observations. Will be recycled if shorter than nclasses.}
}
\value{
A Dataframe with nparam many columns with gaussian data and a columns 'class'
}
\description{
Generates a Dataframe with ninputs many columns and a 'class' column. Every Class has
uncorrelated parameters around a random center within the given cube
boundaries. The Output can be used to create a data_set.
}
\examples{
test <- make_test()
set <- make_set(test, by = class, title="test")

make_test(ninputs = 80, nparam = 4, nclasses = 4, sigma = c(1,2,0.8,1.5), cube = c(-10,10))
}
