% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{make_testset}
\alias{make_testset}
\title{make_testset}
\usage{
make_testset(N = 10, K = 3, P = 2)
}
\arguments{
\item{N}{number of observations per class}

\item{K}{number of classes}

\item{P}{number of parameters of each observation}
}
\value{
a data_set
}
\description{
Creates a simple data_set with random data.
}
\examples{
make_testset(N = 50, K= 2)
}
