% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moore_penrose.R
\name{moore_penrose}
\alias{moore_penrose}
\title{moore_penrose}
\usage{
moore_penrose(mat)
}
\arguments{
\item{mat}{A symmetric, positiv semidefinte Matrix}
}
\value{
The Moore-Penrose Inverse of the given Matrix
}
\description{
moore_penrose
}
