% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_funs.R
\name{PDA}
\alias{PDA}
\title{PDA}
\usage{
PDA(set, base = "id", omega)
}
\arguments{
\item{set}{A R6 data_set object initialized with make_set. @seealso make_set}

\item{base}{One of the following strings \itemize{\item "id"; \item "quad"; \item "cube"; \item "sqrt"; \item "log"; \item "abs"}
The data gets then expanded. @seealso basis_exp}

\item{omega}{A penalizer matrix used for the classification. Note that the dimensions must fit the dimension of the
(potentially) expanded dataset}
}
\value{
Returns a list with the name of the created PDA function in the given set in the first entry and the actual classification
function in the second entry and also saves it in the input set.
}
\description{
The PDA classification function as described in Hastie et al. "The Elements Statistical Learning" (2009)
}
\examples{
test <- make_testset()
func_name <- PDA(test,base='quad')[['name']]
}
