% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_summary}
\alias{plot_summary}
\title{plot_summary}
\usage{
plot_summary(set, name, ppu = 1, background = TRUE, project = TRUE)
}
\arguments{
\item{set}{a data_set}

\item{name}{the name of the function to summarize}

\item{ppu}{background points per unit}

\item{background}{logical, decides if the data plot should have a background grid}

\item{project}{logical, decides if maincomponent analyse should be used to make the data plot}
}
\description{
summarizes a function of a data_set in one plot
}
\examples{
set <- make_testset()
name <- LDA(set)[['name']]
plot_summary(set,name)
}
