% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_funs.R
\name{QDA}
\alias{QDA}
\title{QDA}
\usage{
QDA(set)
}
\arguments{
\item{set}{A R6 data_set object initialized with make_set. @seealso make_set}
}
\value{
Returns a list with the name of the created QDA function in the given set in the first entry and the actual classification
function in the second entry
}
\description{
The QDA classification function as described in Hastie et al. "The Elements Statistical Learning" (2009)
}
\examples{
QDA(Rockets_set)
func_name <- QDA(SAC_G1)[['name']]
}
