% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier_funs.R
\name{RDA}
\alias{RDA}
\title{RDA}
\usage{
RDA(set, alpha, gamma)
}
\arguments{
\item{set}{A R6 data_set object initialized with make_set. @seealso make_set}

\item{alpha}{alpha from formula in Hastie between 0 and 1}

\item{gamma}{gamma from formula between 0 and 1}
}
\value{
Returns a list with the name of the created RDA function in the given set in the first entry and the actual classification
function in the second entry and saves the classification function in the R6 object R6.
}
\description{
The regular Discriminant analysis from 4.3.1 as described in Hastie et al. "The Elements Statistical Learning" (2009)
}
\examples{
test <- make_testset()
func_name <- RDA(test, alpha = 0, gamma = 1)
}
