% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R
\name{validationErrorRate}
\alias{validationErrorRate}
\title{validationErrorRate}
\usage{
validationErrorRate(data, results, alpha, gamma)
}
\value{
total mean error rate
}
\description{
calculates the mean total error rate of RDA for given alpha, gamma to determin 
 best selection in the cross fitting
}
