% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA_crossvalidation.R
\name{alpha_gamma_crossFit}
\alias{alpha_gamma_crossFit}
\title{alpha_gamma_crossFit}
\usage{
alpha_gamma_crossFit(set, K = 3, N = 5)
}
\arguments{
\item{set}{data_set to be trained on for cross validation}

\item{K}{number of validation sets. Note that though K = 10 is common, it does take far too long in this code}

\item{N}{number of parameters to choose. Note that Omega(crossFit) = N^2}
}
\value{
alpha, gamma
}
\description{
cross validates for the best alpha and gamma
}
