% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R, R/plot_functions.R
\name{calc_error}
\alias{calc_error}
\title{calc_error}
\usage{
calc_error(set, name)

calc_error(set, name)
}
\arguments{
\item{set}{A object of class 'data_set'}

\item{name}{the functionname of the function in set}

\item{data}{Dataframe of Parameters for all Observations}

\item{results}{correct classes}

\item{f}{classification function}
}
\value{
total error rate

A list with 3 entrys. Entry [[1]] and [[2]] are matrixes and contain the values as described in the description.
[[3]] is the propability of missclassification.
}
\description{
calculates the total error rate of of a classifaction function on a dataset

Calculates the following propabilitys for a classificationfunction of a data_set:
[[1]]                                                        #accessed over: 
prop. of: class 'A' to be classified to a class 'B'          # ['B','A']
prop. of: class 'A' to be classified to 'A'                  # ['right','A']
prop. of: class 'A' to be not classified to 'A'              # ['wrong','A']
[[2]]
prop. of: classified to class 'A' and actually was class 'B' #['B','A']
prop. of: classified to class 'A' and also was 'A'           #['right','A']
prop. of: classified to class 'A' but was not 'A'            #['wrong','A']
[[3]]
prop. of: missclassification                                 #['miss']
}
