% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{make_set}
\alias{make_set}
\title{make_set}
\usage{
make_set(data, by, title = "", description = "")
}
\arguments{
\item{data}{a Dataframe}

\item{by}{the name of a column of data to classify by}

\item{title}{optional title of the data}

\item{description}{optional description of the data}
}
\value{
A data_set
}
\description{
\code{make_set} transforms a dataframe into an R6 object of class 'data_set'.
}
