% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{make_testset}
\alias{make_testset}
\title{make_testset}
\usage{
make_testset(N = 10, K = 3)
}
\arguments{
\item{N}{number of observations per class}

\item{K}{number of classes}
}
\value{
a data_set
}
\description{
Creates a simple data_set with random data.
}
\examples{
set <- make_testset(N = 50, K= 2)
}
