% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{make_testset}
\alias{make_testset}
\title{make_testset}
\usage{
make_testset(N = 10, G = 3)
}
\arguments{
\item{N}{number of observations per class}

\item{G}{number of classes}
}
\description{
Creates a simple data_set with random data.
}
