% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA.R, R/RDA_crossvalidation.R
\name{validationErrorRate}
\alias{validationErrorRate}
\title{validationErrorRate}
\usage{
validationErrorRate(data, results, alpha, gamma)

validationErrorRate(data, results, alpha, gamma)
}
\arguments{
\item{results}{correct classes}

\item{alpha}{alpha to be evaluated}

\item{gamma}{gamma to be evaluated}

\item{results}{correct classes}

\item{alpha}{alpha to be evaluated}

\item{gamma}{gamma to be evaluated}
}
\value{
total mean error rate of all validations

total mean error rate of all validations
}
\description{
calculates the mean total error rate of RDA for given alpha, gamma to determin 
 best selection in the cross fitting
#'@param data Dataframe of Parameters for all Observations

calculates the mean total error rate of RDA for given alpha, gamma to determin 
 best selection in the cross fitting
#'@param data Dataframe of Parameters for all Observations
}
